pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-news-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "telethon>=1.36",
    "feedparser>=6.0",
    "httpx>=0.27",
    "openai>=1.50",
    "supabase>=2.5",
    "apscheduler>=3.10,<4",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "tenacity>=8.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pillow>=12.2.0",
]
